// src/models/introJsTourt.model.ts
   export interface IntroGuidedTour {
     title?: string;
     element?: string;
     intro?: string;
     position?: 'right' | 'left' | 'bottom' | 'top';
   }
 export interface IntroToursteps {
   title?: string;
   element?: string;
   intro?: string;
   position?: 'right' | 'left' | 'bottom' | 'top';
  }